dev-requirements.txt
airspeed==0.5.5.dev20160812
arrow==0.12.1
asn1crypto==0.24.0
aws-sam-cli==0.5.0
aws-sam-translator==1.6.0
aws-xray-sdk==0.95
awscli==1.15.77
AWSIoTPythonSDK==1.4.0
backports.functools-lru-cache==1.5
backports.ssl-match-hostname==3.5.0.1
backports.tempfile==1.0
backports.weakref==1.0.post1
binaryornot==0.4.4
boto==2.46.1
boto3==1.7.76
botocore==1.10.76
cachetools==0.8.0
certifi==2018.8.13
cffi==1.11.5
chardet==3.0.4
click==6.7
colorama==0.3.9
configparser==3.5.0
cookiecutter==1.6.0
cookies==2.2.1
coverage==4.0.3
cryptography==2.3
dateparser==0.7.0
decorator==4.3.0
dnslib==0.9.7
docker==3.5.0
docker-pycreds==0.3.0
docopt==0.6.2
docutils==0.14
ecdsa==0.13
elasticsearch==6.2.0
enum34==1.1.6
envs==1.2.6
flake8==3.5.0
flake8-quotes==1.0.0
Flask==0.10.1
Flask-Cors==3.0.3
flask-swagger==0.2.12
funcsigs==1.0.2
functools32==3.2.3.post2
future==0.16.0
futures==3.2.0
idna==2.6
ipaddress==1.0.22
itsdangerous==0.24
Jinja2==2.10
jinja2-time==0.2.0
jmespath==0.9.3
jsondiff==1.1.1
jsonpath-rw==1.4.0
jsonpickle==0.9.6
jsonschema==2.6.0
localstack==0.8.7
localstack-client==0.4
localstack-ext==0.8.6
MarkupSafe==1.0
mccabe==0.6.1
mock==2.0.0
moto-ext==1.3.3
nose==1.3.7
pbr==4.2.0
ply==3.11
poyo==0.4.1
psutil==5.4.3
pyaes==1.6.0
pyaml==17.12.1
pyasn1==0.4.4
pycodestyle==2.3.1
pycparser==2.18
pycryptodomex==3.4.9
pyflakes==1.6.0
pyminifier==2.1
pyOpenSSL==17.0.0
python-coveralls==2.7.0
python-dateutil==2.7.3
python-jose-ext==1.3.2.4
pytz==2018.5
PyYAML==3.12
regex==2018.7.11
requests==2.18.4
requests-aws4auth==0.9
responses==0.9.0
RPi.GPIO==0.6.3
rsa==3.4.2
s3transfer==0.1.13
sh==1.12.14
six==1.10.0
srp-ext==1.0.7.1
subprocess32-ext==3.2.8.2
tzlocal==1.5.1
urllib3==1.22
warrant-ext==0.6.1.1
websocket-client==0.48.0
Werkzeug==0.14.1
whichcraft==0.4.1
wrapt==1.10.11
xmltodict==0.10.2
paho-mqtt==1.5.0
